requirements.txt
elasticsearch==8.11.0
python-dotenv==1.0.0
pandas==2.0.3
